+++ src/db/repositories/user-activities.ts
@@ -50,6 +50,8 @@ export async function selectUserActivities({
             'rating', ${schema.reviews.rating},
             'tmdbId', ${schema.reviews.tmdbId},
             'mediaType', ${schema.reviews.mediaType},
+            'seasonNumber', ${schema.reviews.seasonNumber},
+            'episodeNumber', ${schema.reviews.episodeNumber},
             'author', (
               SELECT json_build_object(
                 'id', ${schema.users.id},
@@ -71,6 +73,8 @@ export async function selectUserActivities({
                 'rating', ${schema.reviews.rating},
                 'tmdbId', ${schema.reviews.tmdbId},
                 'mediaType', ${schema.reviews.mediaType},
+                'seasonNumber', ${schema.reviews.seasonNumber},
+                'episodeNumber', ${schema.reviews.episodeNumber},
                 'author', json_build_object(
                   'id', ${schema.users.id},
                   'username', ${schema.users.username},