+++ src/domain/services/list-item/create-list-item.ts
@@ -3,7 +3,6 @@ import { PgIntegrityConstraintViolation } from '@/db/utils/postgres-errors'
 import postgres from 'postgres'
 import type { InsertListItem } from '../../entities/list-item'
 import { ListNotFoundError } from '../../errors/list-not-found-error'
-import { insertUserActivity } from '@/db/repositories/user-activities'
 
 export async function createListItemService(
   values: InsertListItem,
@@ -11,17 +10,6 @@ export async function createListItemService(
 ) {
   try {
     const [listItem] = await insertListItem(values)
-
-    await insertUserActivity({
-      activityType: 'ADD_ITEM',
-      userId,
-      entityId: listItem.listId,
-      entityType: 'LIST',
-      metadata: {
-        tmdbId: listItem.tmdbId,
-        mediaType: listItem.mediaType,
-      },
-    })
 
     return { listItem }
   } catch (error) {