+++ src/domain/services/user-episodes/create-user-episodes.ts
@@ -1,4 +1,3 @@
-import { insertUserActivity } from '@/db/repositories/user-activities'
 import { insertUserEpisodes } from '@/db/repositories/user-episode'
 import type { InsertUserEpisode } from '@/domain/entities/user-episode'
 
@@ -7,14 +6,5 @@ export async function createUserEpisodesService(
   insertActivity = true
 ) {
   const userEpisodes = await insertUserEpisodes(values)
-
-  if (insertActivity) {
-    await insertUserActivity({
-      activityType: 'WATCH_EPISODE',
-      userId: values[0].userId,
-      metadata: values,
-    })
-  }
-
   return { userEpisodes }
 }