+++ src/http/controllers/review-controller.ts
@@ -4,7 +4,6 @@ import {
   createReviewBodySchema,
   getReviewQuerySchema,
   getReviewsQuerySchema,
-  getReviewSummaryQuerySchema,
   reviewParamsSchema,
   updateReviewBodySchema,
 } from '../schemas/reviews'
@@ -17,7 +16,8 @@ import { getReviewsService } from '@/domain/services/reviews/get-reviews'
 import { updateReviewService } from '@/domain/services/reviews/update-review'
 import { getTMDBDataService } from '@/domain/services/tmdb/get-tmdb-data'
 import { getReviewService } from '@/domain/services/reviews/get-review'
-import { getReviewSummaryService } from '@/domain/services/reviews/get-review-summary'
+import { createUserActivity } from '@/domain/services/user-activities/create-user-activity'
+import { deleteUserActivityByEntityService } from '@/domain/services/user-activities/delete-user-activity'
 
 export async function createReviewController(
   request: FastifyRequest,
@@ -34,6 +34,19 @@ export async function createReviewController(
     return reply.status(result.status).send({ message: result.message })
   }
 
+  await createUserActivity({
+    userId: result.review.userId,
+    activityType: 'CREATE_REVIEW',
+    entityId: result.review.id,
+    entityType: 'REVIEW',
+    metadata: {
+      tmdbId: body.tmdbId,
+      mediaType: body.mediaType,
+      seasonNumber: body.seasonNumber,
+      episodeNumber: body.episodeNumber,
+    },
+  })
+
   return reply.status(201).send({ review: result.review })
 }
 
@@ -41,16 +54,25 @@ export async function getReviewsController(
   request: FastifyRequest,
   reply: FastifyReply
 ) {
-  const { mediaType, tmdbId, userId, limit, interval } =
-    getReviewsQuerySchema.parse(request.query)
+  const {
+    mediaType,
+    tmdbId,
+    userId,
+    limit,
+    interval,
+    seasonNumber,
+    episodeNumber,
+  } = getReviewsQuerySchema.parse(request.query)
 
   const result = await getReviewsService({
     mediaType,
     userId,
-    tmdbId: Number(tmdbId),
-    limit: Number(limit),
     authenticatedUserId: request.user?.id,
     interval,
+    tmdbId: Number(tmdbId),
+    limit: Number(limit),
+    seasonNumber: seasonNumber ? Number(seasonNumber) : undefined,
+    episodeNumber: episodeNumber ? Number(episodeNumber) : undefined,
   })
 
   return reply.status(200).send(result.reviews)
@@ -66,6 +88,13 @@ export async function deleteReviewController(
   if (result instanceof DomainError) {
     return reply.status(result.status).send({ message: result.message })
   }
+
+  await deleteUserActivityByEntityService({
+    activityType: 'CREATE_REVIEW',
+    entityType: 'REVIEW',
+    entityId: result.id,
+    userId: result.userId,
+  })
 
   return reply.status(204).send()
 }
@@ -91,10 +120,10 @@ export async function getDetailedReviewsController(
     getReviewsQuerySchema.parse(request.query)
 
   const result = await getReviewsService({
-    limit: Number(limit),
     orderBy,
     userId,
     interval,
+    limit: Number(limit),
   })
 
   const mergedReviews = await Promise.all(
@@ -116,27 +145,15 @@ export async function getReviewController(
   request: FastifyRequest,
   reply: FastifyReply
 ) {
-  const { mediaType, tmdbId } = getReviewQuerySchema.parse(request.query)
+  const { mediaType, tmdbId, seasonNumber, episodeNumber } =
+    getReviewQuerySchema.parse(request.query)
 
   const result = await getReviewService({
     mediaType,
-    tmdbId: Number(tmdbId),
     userId: request.user.id,
-  })
-
-  return reply.status(200).send(result)
-}
-
-export async function getReviewSummaryController(
-  request: FastifyRequest,
-  reply: FastifyReply,
-  redis: FastifyRedis
-) {
-  const values = getReviewSummaryQuerySchema.parse(request.query)
-
-  const result = await getReviewSummaryService({
-    redis,
-    values,
+    tmdbId: Number(tmdbId),
+    seasonNumber: Number(seasonNumber),
+    episodeNumber: Number(episodeNumber),
   })
 
   return reply.status(200).send(result)