+++ src/http/schemas/user-activities.ts
@@ -75,6 +75,8 @@ export const getUserActivitiesResponseSchema = {
               username: z.string(),
               avatarUrl: z.string().nullable(),
             }),
+            seasonNumber: z.number().nullable(),
+            episodeNumber: z.number().nullable(),
           }),
         }),
         z.object({
@@ -97,6 +99,8 @@ export const getUserActivitiesResponseSchema = {
                 username: z.string(),
                 avatarUrl: z.string().nullable(),
               }),
+              seasonNumber: z.number().nullable(),
+              episodeNumber: z.number().nullable(),
             }),
           }),
         }),